+++ lib/src/utils.dart
@@ -8,6 +8,8 @@ import 'dart:collection';
 import 'dart:convert' show UTF8;
 import 'dart:mirrors';
 import 'dart:io';
+import 'package:path/path.dart' as path;
+
 
 Future<String> httpGet(String url) {
   HttpClient client = new HttpClient();
@@ -136,3 +138,41 @@ List<String> coerceToPathList(filesOrPaths) {
     return '${item}';
   }).toList();
 }
+
+
+/// Takes a list of paths and if an element is a directory it expands it to
+/// the Dart source files contained by this directory, otherwise the element is
+/// added to the result unchanged.
+Set<String> findDartSourceFiles(Iterable<String> paths) {
+  /// Returns `true` if this [fileName] is a Dart file.
+  bool _isDartFileName(String fileName) => fileName.endsWith('.dart');
+
+  /// Returns `true` if this relative path is a hidden directory.
+  bool _isInHiddenDir(String relative) =>
+      path.split(relative).any((part) => part.startsWith("."));
+
+  Set<String> _findDartSourceFiles(Directory directory) {
+    var files = new Set<String>();
+    if (directory.existsSync()) {
+      for (var entry
+          in directory.listSync(recursive: true, followLinks: false)) {
+        if (_isDartFileName(entry.path) && !_isInHiddenDir(entry.path)) {
+          files.add(entry.path);
+        }
+      }
+    }
+    return files;
+  }
+
+  var files = new Set<String>();
+
+  paths.forEach((p) {
+    if (FileSystemEntity.typeSync(p) == FileSystemEntityType.DIRECTORY) {
+      files.addAll(_findDartSourceFiles(new Directory(p)));
+    } else {
+      files.add(p);
+    }
+  });
+  return files;
+}
+